backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
SQLAlchemy==1.4.48
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.9.0
